extension/requirements.txt
amqp==2.1.4
Babel==2.4.0
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.2
boto3==1.4.4
botocore==1.5.32
cached-property==1.3.0
celery==4.0.2
certifi==2017.1.23
click==6.7
Django==1.10.5
django-braces==1.11.0
django-celery-beat==1.0.1
django-oauth-toolkit==0.12.0
django-rest-framework-social-oauth2==1.0.5
djangorestframework==3.5.4
djangorestframework-camel-case==0.2.0
djangorestframework-jwt==1.9.0
docker-py==1.9.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
enum34==1.1.6
ffmpy==0.2.2
Flask==0.12
flower==0.9.1
functools32==3.2.3.post2
futures==3.0.5
google-api-python-client==1.6.2
gunicorn==19.7.1
httplib2==0.10.3
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.4
jmespath==0.9.2
kombu==4.0.2
lxml==3.7.3
Markdown==2.6.8
MarkupSafe==0.23
meld3==1.0.2
oauth2client==4.0.0
oauthlib==2.0.1
pbr==1.10.0
psycopg2==2.7
pyasn1==0.2.2
pyasn1-modules==0.0.8
PyJWT==1.4.2
python-dateutil==2.6.0
python-dotenv==0.6.4
python-openid==2.2.5
pytz==2017.2
PyYAML==3.12
requests==2.13.0
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.10
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
stevedore==1.17.1
supervisor==3.3.1
tornado==4.2
uritemplate==3.0.0
vine==1.1.3
virtualenv==15.0.3
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
websocket-client==0.40.0
Werkzeug==0.11.15
Whoosh==2.7.4
youtube-dl==2017.3.26

pydevd
